+++ README.md
@@ -8,6 +8,7 @@
 
 ## Домашнее задание 1
 Инициализация проекта на Gradle
+
 Цель: 
 - создать проект на Котлин;
 - практически разобраться в Gradle Kotlin DSL;
@@ -24,6 +25,7 @@
 
 ## Домашнее задание 2
 Сделать транспортные модели для выпускного проекта с использованием одной из рассмотренных в модуле технологий
+
 Цель: 
 - освоить ООП в Котлин;
 - создать внутренние и контекстные модели;
@@ -38,6 +40,7 @@
 
 ## Домашнее задание 3
 Сделать модуль фреймворка для выпускного проекта
+
 Цель: 
 - освоить работу с одним из фреймворков: Spring Framework, KTOR, Kotless, либо какой-то другой по вашему выбору;
 - научиться работать с шаблоном Dependency Injection;
@@ -50,6 +53,7 @@
 
 ## Домашнее задание 4
 Создание модуля бизнес-логики
+
 Цель: 
 - изучить шаблон проектирования chain of responsibility;
 - создать модуль бизнес-логики;
@@ -62,3 +66,16 @@
 - создать автоматические тесты для различных кейсов для тестирования заглушек;
 - выполнить валидацию входящих запросов;
 - сделать автотесты для валидации.
+
+## Домашнее задание 5
+Обработка асинхронных запросов
+
+Цель:
+- изучить особенности асинхронных фреймворков: по выбору Websocket, RabbitMQ, Kafka;
+- научиться запускать сервера сообщений в docker-контейнерах;
+- реализовать асинхронное API в выпускном проекте с использованием одного из фреймворковх
+- выбрать асинхронный фреймворк;
+- реализовать контроллер для приема асинхронных сообщений;
+- подключить модуль бизнес-логики к контроллеру;
+- обеспечить отправку в сервер сообщений;
+- выполнить тестирование с TestContainers.
