+++ build.gradle.kts
@@ -37,7 +37,7 @@ subprojects {
     tasks {
         withType<KotlinCompile> {
             kotlinOptions {
-                freeCompilerArgs = listOf("-Xjsr305=strict")
+                freeCompilerArgs = listOf("-Xjsr305=strict","-Xopt-in=kotlin.RequiresOptIn")
                 jvmTarget = javaVersion
             }
         }
