+++ gradle.properties
@@ -14,7 +14,7 @@ springDependencyVersion=1.0.11.RELEASE
 springBootVersion=2.4.1
 springFuVersion=0.4.3
 
-ktorVersion=1.5.0
+ktorVersion=1.5.3
 kotlessVersion=0.1.7-beta-5
 
 cassandraVersion=3.10.2
