--- ok-propertysale-be-app-ktor/src/main/kotlin/ru/otus/otuskotlin/propertysale/be/app/ktor/PsBeAppKtor.kt
@@ -1,3 +1,0 @@
-package ru.otus.otuskotlin.propertysale.be.app.ktor
-
-fun main(args: Array<String>): Unit = io.ktor.server.netty.EngineMain.main(args)
