+++ ok-propertysale-be-app-ktor/src/main/kotlin/ru/otus/otuskotlin/propertysale/be/app/ktor/controllers/FlatController.kt
@@ -1,14 +1,8 @@
 package ru.otus.otuskotlin.propertysale.be.app.ktor.controllers
 
 import io.ktor.routing.*
+import ru.otus.otuskotlin.propertysale.be.app.ktor.services.FlatService
 import ru.otus.otuskotlin.propertysale.be.app.ktor.utils.handleRoute
-import ru.otus.otuskotlin.propertysale.be.business.logic.FlatCrud
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondFlatCreate
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondFlatDelete
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondFlatList
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondFlatRead
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondFlatUpdate
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.setQuery
 import ru.otus.otuskotlin.propertysale.mp.common.RestEndpoints
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.requests.PsRequestFlatCreate
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.requests.PsRequestFlatDelete
@@ -21,40 +15,30 @@ import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.responses.PsResponse
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.responses.PsResponseFlatRead
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.responses.PsResponseFlatUpdate
 
-fun Routing.flatRouting(crud: FlatCrud) {
+fun Routing.flatRouting(service: FlatService) {
     post(RestEndpoints.flatList) {
         handleRoute<PsRequestFlatList, PsResponseFlatList> { query ->
-            query?.also { setQuery(it) }
-            crud.list(this)
-            respondFlatList()
+            service.list(this, query)
         }
     }
     post(RestEndpoints.flatCreate) {
         handleRoute<PsRequestFlatCreate, PsResponseFlatCreate> { query ->
-            query?.also { setQuery(it) }
-            crud.create(this)
-            respondFlatCreate()
+            service.create(this, query)
         }
     }
     post(RestEndpoints.flatRead) {
         handleRoute<PsRequestFlatRead, PsResponseFlatRead> { query ->
-            query?.also { setQuery(it) }
-            crud.read(this)
-            respondFlatRead()
+            service.read(this, query)
         }
     }
     post(RestEndpoints.flatUpdate) {
         handleRoute<PsRequestFlatUpdate, PsResponseFlatUpdate> { query ->
-            query?.also { setQuery(it) }
-            crud.update(this)
-            respondFlatUpdate()
+            service.update(this, query)
         }
     }
     post(RestEndpoints.flatDelete) {
         handleRoute<PsRequestFlatDelete, PsResponseFlatDelete> { query ->
-            query?.also { setQuery(it) }
-            crud.delete(this)
-            respondFlatDelete()
+            service.delete(this, query)
         }
     }
 }
