+++ ok-propertysale-be-app-ktor/src/main/kotlin/ru/otus/otuskotlin/propertysale/be/app/ktor/controllers/HouseController.kt
@@ -1,14 +1,8 @@
 package ru.otus.otuskotlin.propertysale.be.app.ktor.controllers
 
 import io.ktor.routing.*
+import ru.otus.otuskotlin.propertysale.be.app.ktor.services.HouseService
 import ru.otus.otuskotlin.propertysale.be.app.ktor.utils.handleRoute
-import ru.otus.otuskotlin.propertysale.be.business.logic.HouseCrud
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondHouseCreate
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondHouseDelete
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondHouseList
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondHouseRead
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondHouseUpdate
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.setQuery
 import ru.otus.otuskotlin.propertysale.mp.common.RestEndpoints
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.requests.PsRequestHouseCreate
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.requests.PsRequestHouseDelete
@@ -21,40 +15,30 @@ import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.responses.PsRespons
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.responses.PsResponseHouseRead
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.responses.PsResponseHouseUpdate
 
-fun Routing.houseRouting(crud: HouseCrud) {
+fun Routing.houseRouting(service: HouseService) {
     post(RestEndpoints.houseList) {
         handleRoute<PsRequestHouseList, PsResponseHouseList> { query ->
-            query?.also { setQuery(it) }
-            crud.list(this)
-            respondHouseList()
+            service.list(this, query)
         }
     }
     post(RestEndpoints.houseCreate) {
         handleRoute<PsRequestHouseCreate, PsResponseHouseCreate> { query ->
-            query?.also { setQuery(it) }
-            crud.create(this)
-            respondHouseCreate()
+            service.create(this, query)
         }
     }
     post(RestEndpoints.houseRead) {
         handleRoute<PsRequestHouseRead, PsResponseHouseRead> { query ->
-            query?.also { setQuery(it) }
-            crud.read(this)
-            respondHouseRead()
+            service.read(this, query)
         }
     }
     post(RestEndpoints.houseUpdate) {
         handleRoute<PsRequestHouseUpdate, PsResponseHouseUpdate> { query ->
-            query?.also { setQuery(it) }
-            crud.update(this)
-            respondHouseUpdate()
+            service.update(this, query)
         }
     }
     post(RestEndpoints.houseDelete) {
         handleRoute<PsRequestHouseDelete, PsResponseHouseDelete> { query ->
-            query?.also { setQuery(it) }
-            crud.delete(this)
-            respondHouseDelete()
+            service.delete(this, query)
         }
     }
 }
