+++ ok-propertysale-be-app-ktor/src/main/kotlin/ru/otus/otuskotlin/propertysale/be/app/ktor/controllers/RoomController.kt
@@ -1,14 +1,8 @@
 package ru.otus.otuskotlin.propertysale.be.app.ktor.controllers
 
 import io.ktor.routing.*
+import ru.otus.otuskotlin.propertysale.be.app.ktor.services.RoomService
 import ru.otus.otuskotlin.propertysale.be.app.ktor.utils.handleRoute
-import ru.otus.otuskotlin.propertysale.be.business.logic.RoomCrud
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondRoomCreate
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondRoomDelete
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondRoomList
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondRoomRead
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.respondRoomUpdate
-import ru.otus.otuskotlin.propertysale.be.mappers.ps.setQuery
 import ru.otus.otuskotlin.propertysale.mp.common.RestEndpoints
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.requests.PsRequestRoomCreate
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.requests.PsRequestRoomDelete
@@ -21,40 +15,30 @@ import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.responses.PsResponse
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.responses.PsResponseRoomRead
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.responses.PsResponseRoomUpdate
 
-fun Routing.roomRouting(crud: RoomCrud) {
+fun Routing.roomRouting(service: RoomService) {
     post(RestEndpoints.roomList) {
         handleRoute<PsRequestRoomList, PsResponseRoomList> { query ->
-            query?.also { setQuery(it) }
-            crud.list(this)
-            respondRoomList()
+            service.list(this, query)
         }
     }
     post(RestEndpoints.roomCreate) {
         handleRoute<PsRequestRoomCreate, PsResponseRoomCreate> { query ->
-            query?.also { setQuery(it) }
-            crud.create(this)
-            respondRoomCreate()
+            service.create(this, query)
         }
     }
     post(RestEndpoints.roomRead) {
         handleRoute<PsRequestRoomRead, PsResponseRoomRead> { query ->
-            query?.also { setQuery(it) }
-            crud.read(this)
-            respondRoomRead()
+            service.read(this, query)
         }
     }
     post(RestEndpoints.roomUpdate) {
         handleRoute<PsRequestRoomUpdate, PsResponseRoomUpdate> { query ->
-            query?.also { setQuery(it) }
-            crud.update(this)
-            respondRoomUpdate()
+            service.update(this, query)
         }
     }
     post(RestEndpoints.roomDelete) {
         handleRoute<PsRequestRoomDelete, PsResponseRoomDelete> { query ->
-            query?.also { setQuery(it) }
-            crud.delete(this)
-            respondRoomDelete()
+            service.delete(this, query)
         }
     }
 }
