+++ ok-propertysale-be-app-ktor/src/main/resources/application.conf
@@ -4,8 +4,6 @@ ktor {
         port = ${?PORT}
     }
     application {
-        modules = [
-            ru.otus.otuskotlin.propertysale.be.app.ktor.AppModuleKt.commonModule,
-       ]
+        modules = [ ru.otus.otuskotlin.propertysale.be.app.ktor.ApplicationKt.module ]
     }
 }
