+++ ok-propertysale-be-app-ktor/src/test/kotlin/ru/otus/otuskotlin/propertysale/be/app/ktor/validaton/flat/FlatUpdateValidationTest.kt
@@ -1,32 +1,45 @@
-package ru.otus.otuskotlin.propertysale.be.app.ktor.flat
+package ru.otus.otuskotlin.propertysale.be.app.ktor.validaton.flat
 
 import io.ktor.http.*
 import io.ktor.server.testing.*
 import ru.otus.otuskotlin.propertysale.be.app.ktor.config.jsonConfig
 import ru.otus.otuskotlin.propertysale.be.app.ktor.module
 import ru.otus.otuskotlin.propertysale.mp.common.RestEndpoints
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.models.PsActionDto
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.PsMessage
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.PsWorkModeDto
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.ResponseStatusDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.requests.PsRequestFlatDelete
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.responses.PsResponseFlatDelete
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.models.PsFlatUpdateDto
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.requests.PsRequestFlatUpdate
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.responses.PsResponseFlatUpdate
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import kotlin.test.fail
 
-class FlatDeleteValidationTest {
+class FlatUpdateValidationTest {
 
     @Test
-    fun `non-empty delete must success`() {
+    fun `non-empty update must success`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.flatDelete) {
-                val body = PsRequestFlatDelete(
+            handleRequest(HttpMethod.Post, RestEndpoints.flatUpdate) {
+                val body = PsRequestFlatUpdate(
                     requestId = "test-request-id",
-                    flatId = "test-flat-id",
-                    debug = PsRequestFlatDelete.Debug(
+                    updateData = PsFlatUpdateDto(
+                        id = "flat-test-id",
+                        name = "flat-test-name",
+                        description = "flat-test-description",
+                        floor = 5,
+                        numberOfRooms = 2,
+                        actions = setOf(
+                            PsActionDto("test-action-1"),
+                            PsActionDto("test-action-2"),
+                            PsActionDto("test-action-3")
+                        )
+                    ),
+                    debug = PsRequestFlatUpdate.Debug(
                         mode = PsWorkModeDto.TEST,
-                        stubCase = PsRequestFlatDelete.StubCase.SUCCESS
+                        stubCase = PsRequestFlatUpdate.StubCase.SUCCESS
                     )
                 )
 
@@ -40,7 +53,7 @@ class FlatDeleteValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatDelete)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatUpdate)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.SUCCESS, res.status)
@@ -52,11 +65,12 @@ class FlatDeleteValidationTest {
     }
 
     @Test
-    fun `empty flat id must fail`() {
+    fun `empty id or title or description must fail`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.flatDelete) {
-                val body = PsRequestFlatDelete(
+            handleRequest(HttpMethod.Post, RestEndpoints.flatUpdate) {
+                val body = PsRequestFlatUpdate(
                     requestId = "test-request-id",
+                    updateData = PsFlatUpdateDto()
                 )
 
                 val format = jsonConfig
@@ -70,15 +84,21 @@ class FlatDeleteValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatDelete)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatUpdate)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.BAD_REQUEST, res.status)
                 assertEquals("test-request-id", res.onRequest)
-                assertTrue("errors: ${res.errors}") {
+                assertTrue {
                     res.errors?.firstOrNull {
-                        it.message?.toLowerCase()?.contains("id") == true
-                            && it.message?.toLowerCase()?.contains("empty") == true
+                        it.message?.contains("name") == true
+                            && it.message?.contains("empty") == true
+                    } != null
+                }
+                assertTrue {
+                    res.errors?.firstOrNull {
+                        it.message?.contains("description") == true
+                            && it.message?.contains("empty") == true
                     } != null
                 }
             }
@@ -88,7 +108,7 @@ class FlatDeleteValidationTest {
     @Test
     fun `bad json must fail`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.flatDelete) {
+            handleRequest(HttpMethod.Post, RestEndpoints.flatUpdate) {
                 val bodyString = "{"
                 setBody(bodyString)
                 addHeader("Content-Type", "application/json")
@@ -98,7 +118,7 @@ class FlatDeleteValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatDelete)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatUpdate)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.BAD_REQUEST, res.status)
