+++ ok-propertysale-be-app-ktor/src/test/kotlin/ru/otus/otuskotlin/propertysale/be/app/ktor/validaton/house/HouseListValidationTest.kt
@@ -1,4 +1,4 @@
-package ru.otus.otuskotlin.propertysale.be.app.ktor.room
+package ru.otus.otuskotlin.propertysale.be.app.ktor.validaton.house
 
 import io.ktor.http.*
 import io.ktor.server.testing.*
@@ -8,28 +8,28 @@ import ru.otus.otuskotlin.propertysale.mp.common.RestEndpoints
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.PsMessage
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.PsWorkModeDto
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.ResponseStatusDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.models.PsRoomListFilterDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.requests.PsRequestRoomList
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.responses.PsResponseRoomList
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.models.PsHouseListFilterDto
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.requests.PsRequestHouseList
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.responses.PsResponseHouseList
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import kotlin.test.fail
 
-class RoomListValidationTest {
+class HouseListValidationTest {
 
     @Test
     fun `non-empty list must success`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.roomList) {
-                val body = PsRequestRoomList(
+            handleRequest(HttpMethod.Post, RestEndpoints.houseList) {
+                val body = PsRequestHouseList(
                     requestId = "test-request-id",
-                    filterData = PsRoomListFilterDto(
+                    filterData = PsHouseListFilterDto(
 
                     ),
-                    debug = PsRequestRoomList.Debug(
+                    debug = PsRequestHouseList.Debug(
                         mode = PsWorkModeDto.TEST,
-                        stubCase = PsRequestRoomList.StubCase.SUCCESS
+                        stubCase = PsRequestHouseList.StubCase.SUCCESS
                     )
                 )
 
@@ -43,12 +43,12 @@ class RoomListValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseRoomList)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseHouseList)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.SUCCESS, res.status)
                 assertEquals("test-request-id", res.onRequest)
-                assertEquals("room-test-name", res.rooms?.firstOrNull()?.name)
+                assertEquals("house-test-name", res.houses?.firstOrNull()?.name)
             }
         }
     }
@@ -56,7 +56,7 @@ class RoomListValidationTest {
     @Test
     fun `bad json must fail`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.roomList) {
+            handleRequest(HttpMethod.Post, RestEndpoints.houseList) {
                 val bodyString = "{"
                 setBody(bodyString)
                 addHeader("Content-Type", "application/json")
@@ -66,7 +66,7 @@ class RoomListValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseRoomList)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseHouseList)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.BAD_REQUEST, res.status)
