+++ ok-propertysale-be-app-ktor/src/test/kotlin/ru/otus/otuskotlin/propertysale/be/app/ktor/validaton/room/RoomReadValidationTest.kt
@@ -1,72 +1,29 @@
-package ru.otus.otuskotlin.propertysale.be.app.ktor.house
+package ru.otus.otuskotlin.propertysale.be.app.ktor.validaton.room
 
 import io.ktor.http.*
 import io.ktor.server.testing.*
 import ru.otus.otuskotlin.propertysale.be.app.ktor.config.jsonConfig
 import ru.otus.otuskotlin.propertysale.be.app.ktor.module
 import ru.otus.otuskotlin.propertysale.mp.common.RestEndpoints
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.models.PsActionDto
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.PsMessage
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.PsWorkModeDto
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.ResponseStatusDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.models.PsHouseCreateDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.requests.PsRequestHouseCreate
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.house.responses.PsResponseHouseCreate
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.requests.PsRequestRoomRead
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.responses.PsResponseRoomRead
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import kotlin.test.fail
 
-class HouseCreateValidationTest {
+class RoomReadValidationTest {
 
     @Test
-    fun `non-empty create must success`() {
+    fun `non-empty roomId must success`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.houseCreate) {
-                val body = PsRequestHouseCreate(
+            handleRequest(HttpMethod.Post, RestEndpoints.roomRead) {
+                val body = PsRequestRoomRead(
                     requestId = "test-request-id",
-                    createData = PsHouseCreateDto(
-                        name = "house-test-name",
-                        description = "house-test-description",
-                        area = 150.0,
-                        actions = setOf(
-                            PsActionDto("test-action-1")
-                        )
-                    ),
-                    debug = PsRequestHouseCreate.Debug(
-                        mode = PsWorkModeDto.TEST,
-                        stubCase = PsRequestHouseCreate.StubCase.SUCCESS
-                    )
-                )
-
-                val bodyString = jsonConfig.encodeToString(PsMessage.serializer(), body)
-                println("REQUEST JSON: $bodyString")
-                setBody(bodyString)
-                addHeader("Content-Type", "application/json")
-            }.apply {
-                assertEquals(HttpStatusCode.OK, response.status())
-                assertEquals(ContentType.Application.Json.withCharset(Charsets.UTF_8), response.contentType())
-                val jsonString = response.content ?: fail("Null response json")
-                println(jsonString)
-
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseHouseCreate)
-                    ?: fail("Incorrect response format")
-
-                assertEquals(ResponseStatusDto.SUCCESS, res.status)
-                assertEquals("test-request-id", res.onRequest)
-                assertEquals("house-test-name", res.house?.name)
-            }
-        }
-    }
-
-    @Test
-    fun `empty title or description must fail`() {
-        withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.houseCreate) {
-                val body = PsRequestHouseCreate(
-                    requestId = "test-request-id",
-                    createData = PsHouseCreateDto(
-                    )
+                    roomId = "test-room-id",
+                    debug = PsRequestRoomRead.Debug(stubCase = PsRequestRoomRead.StubCase.SUCCESS)
                 )
 
                 val format = jsonConfig
@@ -80,23 +37,41 @@ class HouseCreateValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseHouseCreate)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseRoomRead)
+                    ?: fail("Incorrect response format")
+
+                assertEquals(ResponseStatusDto.SUCCESS, res.status)
+                assertEquals("test-request-id", res.onRequest)
+                assertEquals("room-test-name", res.room?.name)
+            }
+        }
+    }
+
+    @Test
+    fun `empty roomId must fail`() {
+        withTestApplication({ module(testing = true) }) {
+            handleRequest(HttpMethod.Post, RestEndpoints.roomRead) {
+                val body = PsRequestRoomRead(
+                    requestId = "test-request-id",
+                    roomId = "",
+                )
+
+                val format = jsonConfig
+
+                val bodyString = format.encodeToString(PsMessage.serializer(), body)
+                setBody(bodyString)
+                addHeader("Content-Type", "application/json")
+            }.apply {
+                assertEquals(HttpStatusCode.OK, response.status())
+                assertEquals(ContentType.Application.Json.withCharset(Charsets.UTF_8), response.contentType())
+                val jsonString = response.content ?: fail("Null response json")
+                println(jsonString)
+
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseRoomRead)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.BAD_REQUEST, res.status)
                 assertEquals("test-request-id", res.onRequest)
-                assertTrue {
-                    res.errors?.firstOrNull {
-                        it.message?.contains("name") == true
-                            && it.message?.contains("empty") == true
-                    } != null
-                }
-                assertTrue {
-                    res.errors?.firstOrNull {
-                        it.message?.contains("description") == true
-                            && it.message?.contains("empty") == true
-                    } != null
-                }
             }
         }
     }
@@ -104,7 +79,7 @@ class HouseCreateValidationTest {
     @Test
     fun `bad json must fail`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.houseCreate) {
+            handleRequest(HttpMethod.Post, RestEndpoints.roomRead) {
                 val bodyString = "{"
                 setBody(bodyString)
                 addHeader("Content-Type", "application/json")
@@ -114,7 +89,7 @@ class HouseCreateValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseHouseCreate)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseRoomRead)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.BAD_REQUEST, res.status)
