+++ ok-propertysale-be-app-ktor/src/test/kotlin/ru/otus/otuskotlin/propertysale/be/app/ktor/validaton/room/RoomUpdateValidationTest.kt
@@ -1,34 +1,49 @@
-package ru.otus.otuskotlin.propertysale.be.app.ktor.flat
+package ru.otus.otuskotlin.propertysale.be.app.ktor.validaton.room
 
 import io.ktor.http.*
 import io.ktor.server.testing.*
 import ru.otus.otuskotlin.propertysale.be.app.ktor.config.jsonConfig
 import ru.otus.otuskotlin.propertysale.be.app.ktor.module
 import ru.otus.otuskotlin.propertysale.mp.common.RestEndpoints
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.models.PsActionDto
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.PsMessage
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.PsWorkModeDto
 import ru.otus.otuskotlin.propertysale.mp.transport.ps.common.transport.ResponseStatusDto
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.requests.PsRequestFlatRead
-import ru.otus.otuskotlin.propertysale.mp.transport.ps.flat.responses.PsResponseFlatRead
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.models.PsRoomUpdateDto
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.requests.PsRequestRoomUpdate
+import ru.otus.otuskotlin.propertysale.mp.transport.ps.room.responses.PsResponseRoomUpdate
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 import kotlin.test.fail
 
-class FlatReadValidationTest {
+class RoomUpdateValidationTest {
 
     @Test
-    fun `non-empty flatId must success`() {
+    fun `non-empty update must success`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.flatRead) {
-                val body = PsRequestFlatRead(
+            handleRequest(HttpMethod.Post, RestEndpoints.roomUpdate) {
+                val body = PsRequestRoomUpdate(
                     requestId = "test-request-id",
-                    flatId = "test-flat-id",
-                    debug = PsRequestFlatRead.Debug(stubCase = PsRequestFlatRead.StubCase.SUCCESS)
+                    updateData = PsRoomUpdateDto(
+                        id = "room-test-id",
+                        name = "room-test-name",
+                        description = "room-test-description",
+                        length = 7.0,
+                        width = 5.0,
+                        actions = setOf(
+                            PsActionDto("test-action-1"),
+                            PsActionDto("test-action-2")
+                        )
+                    ),
+                    debug = PsRequestRoomUpdate.Debug(
+                        mode = PsWorkModeDto.TEST,
+                        stubCase = PsRequestRoomUpdate.StubCase.SUCCESS
+                    )
                 )
 
-                val format = jsonConfig
-
-                val bodyString = format.encodeToString(PsMessage.serializer(), body)
+                val bodyString = jsonConfig.encodeToString(PsMessage.serializer(), body)
+                println("REQUEST JSON: $bodyString")
                 setBody(bodyString)
                 addHeader("Content-Type", "application/json")
             }.apply {
@@ -37,23 +52,24 @@ class FlatReadValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatRead)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseRoomUpdate)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.SUCCESS, res.status)
                 assertEquals("test-request-id", res.onRequest)
-                assertEquals("flat-test-name", res.flat?.name)
+                assertEquals("room-test-id", res.room?.id)
+                assertEquals("room-test-name", res.room?.name)
             }
         }
     }
 
     @Test
-    fun `empty flatId must fail`() {
+    fun `empty id or title or description must fail`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.flatRead) {
-                val body = PsRequestFlatRead(
+            handleRequest(HttpMethod.Post, RestEndpoints.roomUpdate) {
+                val body = PsRequestRoomUpdate(
                     requestId = "test-request-id",
-                    flatId = "",
+                    updateData = PsRoomUpdateDto()
                 )
 
                 val format = jsonConfig
@@ -67,11 +83,23 @@ class FlatReadValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatRead)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseRoomUpdate)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.BAD_REQUEST, res.status)
                 assertEquals("test-request-id", res.onRequest)
+                assertTrue {
+                    res.errors?.firstOrNull {
+                        it.message?.contains("name") == true
+                            && it.message?.contains("empty") == true
+                    } != null
+                }
+                assertTrue {
+                    res.errors?.firstOrNull {
+                        it.message?.contains("description") == true
+                            && it.message?.contains("empty") == true
+                    } != null
+                }
             }
         }
     }
@@ -79,7 +107,7 @@ class FlatReadValidationTest {
     @Test
     fun `bad json must fail`() {
         withTestApplication({ module(testing = true) }) {
-            handleRequest(HttpMethod.Post, RestEndpoints.flatRead) {
+            handleRequest(HttpMethod.Post, RestEndpoints.roomUpdate) {
                 val bodyString = "{"
                 setBody(bodyString)
                 addHeader("Content-Type", "application/json")
@@ -89,7 +117,7 @@ class FlatReadValidationTest {
                 val jsonString = response.content ?: fail("Null response json")
                 println(jsonString)
 
-                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseFlatRead)
+                val res = (jsonConfig.decodeFromString(PsMessage.serializer(), jsonString) as? PsResponseRoomUpdate)
                     ?: fail("Incorrect response format")
 
                 assertEquals(ResponseStatusDto.BAD_REQUEST, res.status)
