+++ ok-propertysale-be-common/src/main/kotlin/ru/otus/otuskotlin/propertysale/be/common/repositories/EmptyUserSession.kt
@@ -1,0 +1,5 @@
+package ru.otus.otuskotlin.propertysale.be.common.repositories
+
+object EmptyUserSession : IUserSession<Any> {
+    override val fwSession = object {}
+}
