+++ ok-propertysale-be-common/src/main/kotlin/ru/otus/otuskotlin/propertysale/be/common/repositories/IUserSession.kt
@@ -1,0 +1,5 @@
+package ru.otus.otuskotlin.propertysale.be.common.repositories
+
+interface IUserSession<T> {
+    val fwSession: T
+}
